various sysfile
the gpio kernel driver now uses a char node in dev instead of files in
/sys
leds sys files are scathered all over /sys/devices and there are only
symlinks in sysclass. there is little point in giving this a private
type by default.

create a private type for /sys/class/zramcontrol so that we can
differentiate between that and /sys/devices/virtual/block/zramN

rename zramclasssysfile to zramcontrolclasssysfile